move spherical_offsets_to to SkyCoord
